fix: renamed components, separated logic into hook
